% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ad_data}
\alias{ad_data}
\title{Gene expression data of Alzeihemer's disease patients}
\format{
A data frame with 259 rows and 294 variables:
\describe{
\item{Label}{Braak stage of Alzheimer's disease}
\item{...}{log10-transformed gene counts}
}
}
\source{
\url{https://adknowledgeportal.synapse.org/}
}
\usage{
ad_data
}
\description{
A dataset containing log10-transformed gene expression counts of 293 genes
from the brains of 259 patients with different stages of Alzheimer's disease.
}
\keyword{datasets}
