% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal_ridge.R
\name{evaluate_ranking}
\alias{evaluate_ranking}
\title{Evaluate rank prediction performance}
\usage{
evaluate_ranking(labels, scores)
}
\arguments{
\item{labels}{Ordered factor vector of true labels for each observation}

\item{scores}{Numeric vector of a score ranking the observations in the predicted order}
}
\value{
A score between 0 and 1. Higher scores mean better prediction performance
}
\description{
Evaluate rank prediction performance
}
