% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal_ridge.R
\name{fobj}
\alias{fobj}
\title{Objective function to minimize}
\usage{
fobj(K, y, v, b, lambda = 1)
}
\arguments{
\item{K}{n-by-n kernel matrix}

\item{y}{n-by-1 vector of (ordinal) labels}

\item{v}{kernel weights}

\item{b}{vector of p bias terms, where p = #classes - 1}

\item{lambda}{ridge regularization coefficient}
}
\description{
Objective function to minimize
}
