% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal_ridge.R
\name{ordinalRidge}
\alias{ordinalRidge}
\title{Ordinal regression with a ridge regularization penalty}
\usage{
ordinalRidge(K, y, lambda = 0.1, eps = 1e-05, maxIter = 10, verbose = TRUE)
}
\arguments{
\item{K}{n-by-n kernel matrix}

\item{y}{n-by-1 vector of (ordinal) labels}

\item{lambda}{regularization coefficient}

\item{eps}{convergence tolerance}

\item{maxIter}{maximum number of iterations}

\item{verbose}{if TRUE, reports objective function value at each iteration}
}
\description{
Ordinal regression with a ridge regularization penalty
}
