% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy.R
\name{toyData}
\alias{toyData}
\title{Generates toy data from sequential Gaussian distributions}
\usage{
toyData(n, p, nb = 6, stdev = 1)
}
\arguments{
\item{n}{Desired number of samples in each class. Values are recycled.}

\item{p}{Data dimensionality}

\item{nb}{Number of boundaries (== number of classes - 1)}

\item{stdev}{Standard deviation of each Gaussian. Values are recycled.}
}
\value{
A list with the following elements:
\describe{
\item{X}{A ns-by-p matrix of ns = sum(n) data points in p dimensions}
\item{y}{A matching ordered factor of length n containing true labels}
}
}
\description{
The n\link{i} data points for the i^th class are generated from N(i,stdev\link{i}).
Coordinates are assumed to be indepedent (i.e., covariance is the identity matrix).
}
