% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal_ridge.R
\name{ordinalRidge}
\alias{ordinalRidge}
\alias{predict.ordinalRidge}
\title{Ordinal regression with a ridge regularization penalty}
\usage{
ordinalRidge(
  K,
  y,
  kernel = (nrow(K) == ncol(K)),
  lambda = 0.1,
  eps = 1e-05,
  maxIter = 10,
  verbose = TRUE
)

\method{predict}{ordinalRidge}(mdl, newdata)
}
\arguments{
\item{K}{an n-by-p matrix of n samples in p dimensions or
an n-by-n kernel matrix}

\item{y}{n-by-1 vector of (ordinal) labels}

\item{kernel}{set TRUE if K is a kernel matrix
(Default: TRUE if K is a square matrix, FALSE otherwise)}

\item{lambda}{regularization coefficient}

\item{eps}{convergence tolerance}

\item{maxIter}{maximum number of iterations}

\item{verbose}{if TRUE, reports objective function value at each iteration}

\item{mdl}{model returned by \code{ordinalRidge()}}

\item{newdata}{n1-by-n kernel matrix of n1 new points against n points used for training}
}
\value{
OrdinalRidge returns a list with the following elements:
#' \describe{
\item{v}{A n-by-1 vector of kernel weights}
\item{b}{A nb-by-1 vector of bias terms to be used as decision boundaries}
\item{classes}{Names of classes from the original vector of labels y}
}

predict() returns a list with the following elements:
\describe{
\item{score}{An n1-by-1 vector of scores}
\item{pred}{A factor of length n1 containing predictions}
\item{prob}{An n1-by-nb matrix of probabilities for each of the nb decision boundaries}
}
}
\description{
Ordinal regression with a ridge regularization penalty
}
\section{Methods (by generic)}{
\itemize{
\item \code{predict}: Predict method for ordinalRidge models
}}

